fix: qwen35moe expert mapping + MoE kernel dtype casts
Two bugs blocking end-to-end MoE inference:

1. weights_adapter/default.py: the qwen3_5_moe_text branch was an elif in
   the model-type chain, so it short-circuited before the shared Qwen MoE
   expert mapping could run. Every ffn_{gate,up,down}_exps tensor was
   dropped from the name map, leaving expert weights empty at forward
   (IndexError: index 86 is out of bounds for dimension 0 with size 0).
   Hoisted the normalization above the chain so qwen35moe falls through
   into the shared qwen2_moe/qwen3_moe expert branch.

2. triton/fused_moe/standard_quant/q4_0_rocmfp4{,_fast}.py: the two
   Codebook10 MoE kernels built w_tile from fp32 codebook values scaled
   by an x_dtype scale, which promotes back to fp32, then fed it to
   tl.dot against a bf16 x_tile -> 'Both operands must be the same dtype.
   Got bf16 and fp32'. Added the missing .to(x_dtype) after the reshape,
   matching the pattern the other four ROCmFPX MoE kernels already use.

STRIX_LEAN (Qwen3.5 MoE, 90 experts, type 101) now loads and generates
end-to-end on gfx1151. Output quality not yet validated.

+++ vllm_gguf_plugin/triton/fused_moe/standard_quant/q4_0_rocmfp4.py
@@ -161,7 +161,7 @@ def _q4_0_rocmfp4_moe_kernel(
         w_high = high_codes * high_scale[:, :, None]
         w_tile = tl.reshape(
             tl.join(w_low, w_high), (BLOCK_N, BLOCK_K_BLOCKS * BLOCK_SIZE)
-        )
+        ).to(x_dtype)
 
         acc = tl.dot(x_tile, tl.trans(w_tile), acc=acc)
