+++ vllm_gguf_plugin/triton/fused_moe/standard_quant/q4_0_rocmfp4_fast.py
@@ -155,7 +155,7 @@ def _q4_0_rocmfp4_fast_moe_kernel(
         w_high = high_codes * scale[:, :, None]
         w_tile = tl.reshape(
             tl.join(w_low, w_high), (BLOCK_N, BLOCK_K_BLOCKS * BLOCK_SIZE)
-        )
+        ).to(x_dtype)
 
         acc = tl.dot(x_tile, tl.trans(w_tile), acc=acc)
